Added the GPL v2 license file.
The original source of this repo is https://code.google.com/p/byte-unixbench/
where the project is listed as being released under the GPL v2 license, but
as it was metadata in Google Code but not explicitly named in the project source
code itself, it did not transfer when the project was moved to GitHub.

This change clarifies the license of the project and provides a complete copy of
the license to make it explicit to future users and contributors.