Fix for OS X based on https://gist.github.com/barusan/11033924
barusan's patch mostly retains compatibility with linux, but
unconditionally used machdep instead of /proc/cpuinfo

This attempts to merge the patch without harming behaviour on linux by
detecting the darwin platform and using machdep there but restores
/proc/cpuinfo elsewhere.

The optimisation level is turned down to a minimum in this version
because the default apparently didn't work on OS X which may impact
performance obtained elsewhere. Modifying the makefile by hand will
restore whichever performance level, for example restoring the line
which mentions solaris will put it back to the previous default.